refactor: rename to serialize-dashboard-state
